Created routes.js and moved all routes from app.js into this file. Refactor routes to use express.Router() instead of app.
